% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/juno.R
\name{juno}
\alias{juno}
\title{Display a jupyter notebook file}
\usage{
juno(nbname, port = 8888)
}
\arguments{
\item{port}{the port to display the notebook on localhost}

\item{jupyter_file}{a jupyter notebook file}
}
\description{
Display a jupyter notebook file
}

