% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jupyter_server.R
\name{jupyter_server}
\alias{jupyter_server}
\title{Launch a jupyter server or connect with one already running}
\usage{
jupyter_server(notebook = NULL, dir = NULL, port = 8888,
  ip = "localhost", view = NULL)
}
\arguments{
\item{notebook}{to open. If file does not exist, it will be created, with a \code{.ipynb} extension. Default is none, just launch a server in the background.}

\item{dir}{the server working directory. Default is either the working directory or parent directory of the notebook.}

\item{port}{the port on which to launch the server.  Default is 8888.}

\item{ip}{the ip to launch the server on.  Default is 'localhost'}

\item{view}{open a viewer window or browser with the server? Default is to view only if a notebook is specified.}
}
\description{
Launch a jupyter server or connect with one already running
}

