% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jupyter_server.R
\name{jupyter_server}
\alias{jupyter_server}
\title{Launch a jupyter server or connect with one already running}
\usage{
jupyter_server(notebook = NULL, dir = NULL, port = 8888,
  ip = "localhost", view = NULL)
}
\arguments{
\item{dir}{the server working directory}

\item{port}{the port on which to launch the server}

\item{ip}{the ip to launch the server on}

\item{view}{open a viewer window or browser with the server?}
}
\description{
Launch a jupyter server or connect with one already running
}

