% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jupyter_to_rmd.R
\name{jupyter_to_rmd}
\alias{jupyter_to_rmd}
\title{Convert a Jupyter Notebook file to an Rmd file}
\usage{
jupyter_to_rmd(infile, outfile, view = !is.null(getOption("viewer")),
  header_metadata = TRUE, code_metadata = TRUE, text_metadata = TRUE,
  knitr_metadata = TRUE)
}
\arguments{
\item{infile}{the jupyter file to convert}

\item{outfile}{the target file}
}
\description{
Convert a Jupyter Notebook file to an Rmd file
}

